+++ sources/src/models/GameVector.java
@@ -4,14 +4,14 @@ package models;
  * Created by KhacThanh on 8/6/2016.
  */
 public class GameVector {
-    public int dx;
-    public int dy;
+    public float dx;
+    public float dy;
 
     public GameVector() {
         this(0, 0);
     }
 
-    public GameVector(int dx, int dy) {
+    public GameVector(float dx, float dy) {
         this.dx = dx;
         this.dy = dy;
     }
